requirements.txt

google-api-python-client==2.143.0
google-auth==2.33.0
google-auth-oauthlib==1.2.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.66.4
duckdb==1.0.0
Pillow==10.4.0
opencv-python==4.10.0.84
pytesseract==0.3.13
matplotlib==3.9.0
scikit-learn==1.5.1
sentence-transformers==3.0.1
rank-bm25==0.2.2
